backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
pyyaml==6.0
requests==2.28.2
python-dotenv==1.0.0
marshmallow==3.19.0
gunicorn==20.1.0
pymysql==1.1.0
pyjwt==2.6.0
flasgger==0.9.5
flask-migrate==4.0.4